Modified the acrylic styles to a component to not use fixed background
