+++ src/components/Acrylic.vue
@@ -1,65 +1,66 @@
 <template>
-  <div class="acrylic card">
-    <span class="meter" v-if="meter" :style="{width: meter+'%'}" />
-    <i v-if="icon" :class="'icon '+icon" :style="{color}"></i>
-    <div class="info">
-      <div class="title">{{ title }}</div>
-      <div class="subtitle">{{ subtitle }}</div>
+  <div class="acrylic-container">
+    <div class="child">
+      <slot />
     </div>
+    <div class="acrylic" />
   </div>
 </template>
 
 <script>
 export default {
-  name: 'Acrylic',
-  props: ['icon', 'title', 'subtitle', 'color', 'meter']
+  name: 'Acrylic'
 }
 </script>
 
-<style>
-  .card {
-    width: 100px;
-    height: 100px;
-    display: inline-flex;
-    flex-direction: column;
-    align-items: left;
-    padding: 20px;
-    color: #ffffff;
-    font-weight: 600;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: all 0.3s ease;
+<style scoped>
+  .acrylic-container {
+    position: relative;
+    overflow: hidden;
+    z-index: 0;
   }
-  .card .icon {
-    font-size: 22px;
-    margin-top: 0;
-    margin-bottom: auto;
+
+  .acrylic-container .acrylic {
+    position: absolute !important;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
   }
-  .card .info {
-    margin-bottom: 0;
-    margin-top: auto;
-    text-align: left;
+
+  .acrylic {
+    position: absolute;
+    clip: rect(auto, auto, auto, auto);
+    overflow: hidden;
+    box-shadow: 0 0 30px rgba(46, 46, 46, 0.4);
+    z-index: -1;
   }
-  .card .info .subtitle {
-    font-size: 11px;
-    color: rgba(220, 220, 220, 0.85);
-    margin-top: 10px;
+  .acrylic::before {
+    content: '';
+    position: fixed;
+    top: -15px;
+    left: -15px;
+    right: -15px;
+    bottom: -15px;
+    background: url(https://source.unsplash.com/collection/1501932);
+    /* background-attachment: fixed; */
+    background-size: cover;
+    background-position: center center;
+    z-index: -3;
+    pointer-events: none;
+    filter: blur(10px) saturate(1.2) brightness(0.7);
   }
-  .card .meter {
+  .acrylic::after {
+    content: '';
     position: absolute;
     top: 0;
-    height: 3px;
     left: 0;
-    background-color: orange;
-    opacity: 1;
-    border-bottom-right-radius: 3px;
-    border-top-right-radius: 3px;
+    right: 0;
+    bottom: 0;
+    z-index: -2;
+    pointer-events: none;
+    background-color: rgba(200, 200, 200, 0.28);
+    transition: background-color 0.3s ease;
   }
-  
-  .card:hover {
-    box-shadow: 0 10px 50px rgba(20, 20, 20, 0.7);
-  }
-  .card:hover::after {
-    background-color: rgba(243, 243, 243, 0.45);
-  }
+
 </style>
